sounds, visual stuff

+++ source/bonus/draw_f_and_c/draw_f_and_c.c
@@ -6,7 +6,7 @@
 /*   By: zstenger <zstenger@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/14 16:51:57 by jergashe          #+#    #+#             */
-/*   Updated: 2023/05/15 10:38:29 by zstenger         ###   ########.fr       */
+/*   Updated: 2023/05/15 19:40:20 by zstenger         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,18 +61,18 @@ void	init_variables_for_x(t_map *map, t_fc_tex *fc, int x, int y)
 	fc->tex_y = (int)(64 * (fc->floor_y - fc->cell_y)) & (64 - 1);
 	fc->floor_x += fc->f_step_x;
 	fc->floor_y += fc->f_step_y;
-	color = rgb(fc->tex->pixels[64 * (int)fc->tex_y
-			* 4 + (int)fc->tex_x * 4], \
-		fc->tex->pixels[64 * (int)fc->tex_y * 4 + (int)fc->tex_x * 4 + 1], \
-		fc->tex->pixels[64 * (int)fc->tex_y * 4 + (int)fc->tex_x * 4 + 2], \
-		fc->tex->pixels[64 * (int)fc->tex_y * 4 + (int)fc->tex_x * 4 + 3]);
+	color = rgb((int)fc->tex->pixels[64 * fc->tex_y
+			* 4 + fc->tex_x * 4], \
+		(int)fc->tex->pixels[64 * fc->tex_y * 4 + fc->tex_x * 4 + 1], \
+		(int)fc->tex->pixels[64 * fc->tex_y * 4 + fc->tex_x * 4 + 2], \
+		(int)fc->tex->pixels[64 * fc->tex_y * 4 + fc->tex_x * 4 + 3]);
 	if ((color & 0x00FFFFFF) != 0)
 		map->buffer[x][y] = color;
-	color2 = rgb(fc->tex2->pixels[64 * (int)fc->tex_y
-			* 4 + (int)fc->tex_x * 4], \
-		fc->tex2->pixels[64 * (int)fc->tex_y * 4 + (int)fc->tex_x * 4 + 1], \
-		fc->tex2->pixels[64 * (int)fc->tex_y * 4 + (int)fc->tex_x * 4 + 2], \
-		fc->tex2->pixels[64 * (int)fc->tex_y * 4 + (int)fc->tex_x * 4 + 3]);
+	color2 = rgb((int)fc->tex2->pixels[64 * fc->tex_y
+			* 4 + fc->tex_x * 4], \
+		(int)fc->tex2->pixels[64 * fc->tex_y * 4 + fc->tex_x * 4 + 1], \
+		(int)fc->tex2->pixels[64 * fc->tex_y * 4 + fc->tex_x * 4 + 2], \
+		(int)fc->tex2->pixels[64 * fc->tex_y * 4 + fc->tex_x * 4 + 3]);
 	if ((color2 & 0x00FFFFFF) != 0)
 		map->buffer[x][HEIGHT - y - 1] = color2;
 }
